+++ packages/core/src/cli/cli.test.ts
@@ -607,10 +607,13 @@ test('validate - excludes specified contract', async t => {
   const buildInfo = await artifacts.getBuildInfo(`contracts/test/cli/Validate.sol:BecomesBadLayout`);
   await fs.writeFile(path.join(temp, 'validate.json'), JSON.stringify(buildInfo));
 
-  const error = await t.throwsAsync(execAsync(
-    `${CLI} validate ${temp} --contract BecomesBadLayout --reference StorageV1 --exclude "**/Validate.sol"`,
-  ));
-  t.true(error?.message.includes('No validation report found for contract contracts/test/cli/Validate.sol:BecomesBadLayout'), error?.message);
+  const error = await t.throwsAsync(
+    execAsync(`${CLI} validate ${temp} --contract BecomesBadLayout --reference StorageV1 --exclude "**/Validate.sol"`),
+  );
+  t.true(
+    error?.message.includes('No validation report found for contract contracts/test/cli/Validate.sol:BecomesBadLayout'),
+    error?.message,
+  );
 });
 
 test('validate - excludes one contract from layout comparisions', async t => {
@@ -618,9 +621,7 @@ test('validate - excludes one contract from layout comparisions', async t => {
   const buildInfo = await artifacts.getBuildInfo(`contracts/test/cli/excludes/ImportVersions.sol:Dummy`);
   await fs.writeFile(path.join(temp, 'validate.json'), JSON.stringify(buildInfo));
 
-  const error = await t.throwsAsync(execAsync(
-    `${CLI} validate ${temp} --exclude "**/V2Bad1.sol"`,
-  ));
+  const error = await t.throwsAsync(execAsync(`${CLI} validate ${temp} --exclude "**/V2Bad1.sol"`));
   const expectation: string[] = [`Stdout: ${(error as any).stdout}`, `Stderr: ${(error as any).stderr}`];
   t.snapshot(expectation.join('\n'));
 });